+++ packages/igx-templates/constants.ts
@@ -1,1 +1,1 @@
-export const IGNITEUI_ANGULAR_PACKAGE = "igniteui-angular@~21.0.0";
+export const IGNITEUI_ANGULAR_PACKAGE = "igniteui-angular@~21.1.0";