+++ packages/igx-templates/igx-ts-legacy/autocomplete/autocomplete-extended/index.ts
@@ -18,10 +18,10 @@ class IgxAutocompleteTemplate extends IgniteUIForAngularTemplate {
 			from: "@angular/forms"
 		}, {
 			declare: ["<%=ClassName%>PipeStartsWith"],
-			from: "./src/app/<%=path%>/<%=filePrefix%>.component.ts"
+			from: "./src/app/<%=path%>/<%=filePrefix%>.ts"
 		}, {
 			declare: ["<%=ClassName%>RegionContains"],
-			from: "./src/app/<%=path%>/<%=filePrefix%>.component.ts"
+			from: "./src/app/<%=path%>/<%=filePrefix%>.ts"
 		}];
 	}
 }