+++ packages/igx-templates/igx-ts-legacy/dialog/default/files/src/app/__path__/__filePrefix__.ts
@@ -1,0 +1,15 @@
+import { Component } from '@angular/core';
+import { IDialogEventArgs } from '<%=igxPackage%>';
+
+@Component({
+  selector: 'app-<%=filePrefix%>',
+  templateUrl: './<%=filePrefix%>.html',
+  styleUrl: './<%=filePrefix%>.scss',
+  standalone: false
+})
+
+export class <%=ClassName%> {
+  dialogOKSelected(args: IDialogEventArgs) {
+    args.dialog.close();
+  }
+}