+++ packages/igx-templates/igx-ts-legacy/hierarchical-grid/default/files/src/app/__path__/__filePrefix__.ts
@@ -1,0 +1,12 @@
+import { Component } from '@angular/core';
+import { ARTISTS, Artist } from './data';
+
+@Component({
+  selector: 'app-<%=filePrefix%>',
+  templateUrl: './<%=filePrefix%>.html',
+  styleUrl: './<%=filePrefix%>.scss',
+  standalone: false
+})
+export class <%=ClassName%> {
+  public localData: Artist[] = ARTISTS;
+}