+++ packages/igx-templates/igx-ts-legacy/projects/empty/files/src/app/home/home.ts
@@ -1,0 +1,10 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'app-home',
+  templateUrl: './home.html',
+  styleUrl: './home.scss',
+  standalone: false
+})
+export class Home {
+}