+++ packages/igx-templates/igx-ts-legacy/projects/side-nav-auth/files/src/app/authentication/profile/profile.spec.ts
@@ -1,0 +1,28 @@
+import { ComponentFixture, TestBed } from '@angular/core/testing';
+import { UserService } from '../services/user.service';
+import { Profile } from './profile';
+
+describe('ProfileComponent', () => {
+  let component: Profile;
+  let fixture: ComponentFixture<Profile>;
+
+  beforeEach(async () => {
+    await TestBed.configureTestingModule({
+      declarations: [Profile],
+      providers: [
+        { provide: UserService, useValue: { currentUser: { name: 'test' } } }
+      ]
+    })
+      .compileComponents();
+  });
+
+  beforeEach(() => {
+    fixture = TestBed.createComponent(Profile);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});