+++ packages/igx-templates/igx-ts/hierarchical-grid/default/files/src/app/__path__/__filePrefix__.ts
@@ -1,0 +1,13 @@
+import { Component } from '@angular/core';
+import { ARTISTS, Artist } from './data';
+import { IgxHierarchicalGridComponent, IgxColumnComponent, IgxRowIslandComponent } from '<%=igxPackage%>';
+
+@Component({
+  selector: 'app-<%=filePrefix%>',
+  templateUrl: './<%=filePrefix%>.html',
+  styleUrl: './<%=filePrefix%>.scss',
+  imports: [IgxHierarchicalGridComponent, IgxColumnComponent, IgxRowIslandComponent]
+})
+export class <%=ClassName%> {
+  public localData: Artist[] = ARTISTS;
+}