+++ packages/igx-templates/igx-ts/projects/_base/files/src/app/app.routes.ts
@@ -1,8 +1,8 @@
 import { Routes } from '@angular/router';
-import { PageNotFoundComponent } from './error-routing/not-found/not-found.component';
-import { UncaughtErrorComponent } from './error-routing/error/uncaught-error.component';
+import { NotFound } from './error-routing/not-found/not-found';
+import { UncaughtError } from './error-routing/error/uncaught-error';
 
 export const routes: Routes = [
-  { path: 'error', component: UncaughtErrorComponent },
-  { path: '**', component: PageNotFoundComponent } // must always be last
+  { path: 'error', component: UncaughtError },
+  { path: '**', component: NotFound } // must always be last
 ];