+++ spec/acceptance/generate-spec.ts
@@ -174,10 +174,10 @@ describe("Generate command", () => {
 
 		const pathDirectory = path.join(templateFolderPath, "files", "src", "app", "__path__");
 		expect(fs.existsSync(pathDirectory)).toBeTruthy();
-		expect(fs.existsSync(path.join(pathDirectory, "__filePrefix__.component.scss"))).toBeTruthy();
-		expect(fs.existsSync(path.join(pathDirectory, "__filePrefix__.component.html"))).toBeTruthy();
-		expect(fs.existsSync(path.join(pathDirectory, "__filePrefix__.component.spec.ts"))).toBeTruthy();
-		expect(fs.existsSync(path.join(pathDirectory, "__filePrefix__.component.ts"))).toBeTruthy();
+		expect(fs.existsSync(path.join(pathDirectory, "__filePrefix__.scss"))).toBeTruthy();
+		expect(fs.existsSync(path.join(pathDirectory, "__filePrefix__.html"))).toBeTruthy();
+		expect(fs.existsSync(path.join(pathDirectory, "__filePrefix__.spec.ts"))).toBeTruthy();
+		expect(fs.existsSync(path.join(pathDirectory, "__filePrefix__.ts"))).toBeTruthy();
 	});
 
 	it("Should correctly generate react template", async () => {